requirements.txt
fastmcp>=0.1.0
httpx>=0.25.0
python-dotenv>=1.0.0
uvicorn>=0.24.0